runtime, cmd/ld: use linker magic to remove tls assumptions
The runtime assumes that it knows the offset from the TLS base to g,
but in the external linking case it is actually the host linker that
decides.  This changes the runtime (on linux/amd64 only) to use a
global variable with a magic name that the linker recognizes and
generates the appropriate relocation to overwrite it with the offset.

Fixes #9913

Change-Id: I75050866dbb7f0f5a952ffaa9c4caca6e5212ee7