runtime,cmd/internal/ld: move functab linker symbols into single struct
NOT FOR MERGING

In preparation for being able to run a go program that has code in
several objects, this changes from having several linker symbols used
by the functab/pclntab code into having one linker symbol that points
at a structure containing the data the code needs.  Multiple object
support will construct a linked list of such structures.

It would probably be possible to initialize the slices in the heapsegment
structure directly from the linker but I was aiming for a minimal diff for
now.

Change-Id: I613cce35309801cf265a1d5ae5aaca8d689c5cbf