runtime,cmd/internal/ld,cmd/6l: sketch of support for multiple heap segments
REALLY REALLY NOT FOR MERGING

This builds on the previous commit to show how having multiple
functabs could work.  Obviously much more needs to be done to
actually excercise this code (including care around symbol
visibility), but I have other branches where it seems to work.

It works by having a linked list of heapsegment structures.  The
first entry in the list is the heapsegment for the object containing
the runtime and this is set up by the linker when linking that
object.  Other objects get a function in the .init_array that adds
their heapsegments to the linked list.  I've made minimal changes to
the runtime functions to consult all known segments, it could be done
more efficiently but I wanted a minimal delta for now.

Change-Id: I99ed91e6aa05c633b99d45669c0048b48a903313