cmd/internal/obj/x86, cmd/internal/ld, cmd/6l: support accessing global data via GOT
NOT FOR MERGE

This has caveats: it's overly broad, accessing data via the GOT even
when it's OK to access it directly and it's activated when -shared is
passed to 6g/asm even though it's only useful when Go code is to be
distributed amongst many object files, something that's not yet
possible.

But it's enough to get started with.