cmd/internal/gc: provide value for importpath of package being compiled
Currently the compiler generates a reference to the import path of the
package that is being compiled by emitting to a reference to a symbol
called go.importpath."", which is rewritten by the linker to match the
name of a symbol defined by any package that imports this one.

This doesn't work well in the dynamic linking world, because when an
executable is linked against a shared library there is no reason to
think that it will import (== define go.importpath.* symbols for) all
packages in the shared library, and so you get undefined symbols.

Fortunately the compiler knows the import path of the package being
compiled (if the -p flag is passed, which the go tool always does) so
we can just define the value for the symbol when compiling the
package.

As a bonus, remove some obscure code from the linker.

Change-Id: I57cd5447b42a1a6129b02951d44efffb10cf64be