cmd/internal/obj,etc: replace -shared with -buildmode=c-shared
Currently some compilers, the linker and the assembler take a -shared
flag which means approximately what the -buildmode=c-shared flag means
in the proposed "Go Execution Modes" document. As part of implementing
other modes, the term "shared" becomes horribly overloaded, so this
adds a -buildmode argument instead (which currently only handles
-buildmode=c-shared, no new behaviour here).

The compilers and linker retain a -shared argument as an alias for
-buildmode=c-shared.  The assemblers in 1.4 did not support -shared
and the tool has a different name now in any case, so I don't see a
reason to maintain compatibility there.

Change-Id: I1c36f0060e0c3d256f36d2e9a07990446b756379