reflect,runtime: check equality, not identity, for method names
NOT FOR MERGE

When dynamically linking Go code, it is no longer safe to assume that
strings that end up in method names are identical if they are equal.

This presumably is a pessimisation for exiting Go programs.  I'd love
tips on how to measure how much.

Another approach would be to define a global symbol for all strings
that end up as method names I suppose.

Change-Id: I0b3dfc314215a4d4e09eec6b42c1e3ebce33eb56