cmd/compile, runtime: stop returning type.zero on hash miss
This replaces the scheme where the compiler tracked the largest type that was
the value of a hashtable and allocate a block of zeroes large enough to return
as a hash miss with a scheme that is simpler on the runtime but more
complicated in the compiler: just have a fixed size block of zeroes (256 bytes
here) and handle hash access differently in the compiler when the value type is
too large.

What I have here doesn't quite work (try cranking the <= 256 in walk.go down to
<= 8).

Update #11491.

Change-Id: I1da3aca96bc4e0f32129f3a53e31dd8079f79100