cmd/cgo, cmd/compile, cmd/link: set the ELF headers of ARM executables that use cgo correctly
It is generally expected that the ELF flags of a dynamically linked executable
and the libraries it links against match. Go's linker currently always produces
executables with flags that do not declare a float abi (hard, soft) at all, but
when cgo is involved it is unlikely that this matches the system libraries
being linked against -- really the decision about ABI is made by the C compiler
during the invocation of cgo. So this records, the best way I know how, the abi
that compiler uses during the cgo run and records it in a way that gets through
to the linker, which can then set the flags appropriately.

Fixes #7094

Change-Id: If0d9fc5219549dcc7910c0651017462375e01eb4