cmd/internal/obj, cmd/link: serialize Reloc.Variant in objfile
DO NOT REVIEW. I need to see if this is truly necessary.

This makes it easier to specify different variants of the same
relocation to the linker from liblink.

For example, although most instruction we use for R_ADDRPOWER64
use the same D-Form instruction encoding, floating point load/stores
use the DS-Form, rather than inventing a new relocation time, we
can add a variant to the existing type. The linker already uses
this scheme to reduce code duplication caused by the huge number
of ELF ppc64 relocations.

This change brings the benefit Variant field to liblink as well,
so that we can change 3-instruction constant load to 2-instruction.

Change-Id: Icfae861bd36433cffcc5da0befdeb381c97837ba