cmd/link: force all relocations to symbols to be resolved by system linker when dynamically linking
When dynamically linking, we want references to functions defined
in this module to always be to the function object, not to the
PLT slot so for global function symbols we write an additional
local symbol and make all relocations refer to this local symbol
instead. This is doing what -Bsymbolic-functions for the external
linker is supposed to do but that is buggy on several platforms.

Change-Id: Ie6983eb4d1947f8543736fd349f9a90df3cce91a